Move mod pages to nuxt child (#201)
* Move mod pages to nuxt child

* Minor fixes

* Fix lockfile
