+++ package.json
@@ -22,7 +22,7 @@
     "cookie-universal-nuxt": "^2.1.4",
     "highlight.js": "^10.3.2",
     "marked": "^2.0.0",
-    "nuxt": "^2.14.7",
+    "nuxt": "^2.15.5",
     "v-tooltip": "^2.0.3",
     "vue-click-outside": "^1.1.0",
     "vue-highlightjs": "^1.3.3",
@@ -31,7 +31,7 @@
     "xss": "^1.0.8"
   },
   "devDependencies": {
-    "@nuxt/types": "^2.14.12",
+    "@nuxt/types": "^2.15.5",
     "@nuxtjs/color-mode": "^1.1.1",
     "@nuxtjs/eslint-config": "^3.1.0",
     "@nuxtjs/eslint-module": "^2.0.0",
@@ -43,6 +43,6 @@
     "eslint-plugin-prettier": "^3.1.4",
     "node-sass": "^4.14.1",
     "prettier": "^2.1.2",
-    "sass-loader": "^9.0.3"
+    "sass-loader": "^10.1.1"
   }
 }
